Put a couple mechanisms in place to speed up request handling
* Front-loading all computes/data-retrieval tasks that are a direct function of the request; requests will ideally only do a total of two awaits before returning
* Refactor a few of the follow-request API calls to utilize the above
* Reuse of DynamoDB client

More importantly than the above, I will need to set up an SQS queue to handle all additional API requests asynchronously from the lambda processing. These are very costly requests as they incur the overhead of another HTTPS request to the public internet (we're not using a VPC setup that might minimize this) and of the Cognito authorizer.